+++ model/user.go
@@ -3,6 +3,6 @@ package model
 type User struct {
 	Name        string
 	MailId      string
-	CreditLimit int
-	Dues        int
+	CreditLimit float32
+	Dues        float32
 }
